requirements.txt
alabaster==1.0.0
annotated-types==0.7.0
app-model==0.2.8
appdirs==1.4.4
appnope==0.1.4
asciitree==0.3.3
asttokens==2.4.1
attrs==24.2.0
babel==2.16.0
build==1.2.2
cachey==0.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dask==2024.9.0
debugpy==1.8.5
decorator==5.1.1
docstring_parser==0.16
docutils==0.21.2
exceptiongroup==1.2.2
executing==2.1.0
fasteners==0.19
flexcache==0.3
flexparser==0.3.1
fonttools==4.53.1
freetype-py==2.5.1
fsspec==2024.9.0
HeapDict==1.0.1
hsluv==5.0.4
idna==3.10
imageio==2.35.1
imagesize==1.4.1
in-n-out==0.2.1
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
locket==1.0.0
magicgui==0.9.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
napari==0.5.3
napari-console==0.0.9
napari-plugin-engine==0.2.0
napari-plugin-manager==0.1.1
napari-svg==0.2.0
nest-asyncio==1.6.0
networkx==3.3
npe2==0.7.7
numba==0.60.0
numcodecs==0.13.0
numpy==2.0.2
numpydoc==1.8.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==10.4.0
Pint==0.24.3
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.47
psutil==6.0.0
psygnal==0.11.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-pde==0.41.0
pyconify==0.1.6
pydantic==2.9.2
pydantic-compat==0.1.2
pydantic_core==2.23.4
Pygments==2.18.0
PyOpenGL==3.1.7
pyparsing==3.1.4
pyproject_hooks==1.1.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.6.0
QtPy==2.4.1
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
scikit-image==0.24.0
scipy==1.14.1
setuptools==73.0.1
shellingham==1.5.4
six==1.16.0
snowballstemmer==2.2.0
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
superqt==0.6.7
sympy==1.13.3
tabulate==0.9.0
tifffile==2024.8.30
tomli_w==1.0.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vispy==0.14.3
wcwidth==0.2.13
wheel==0.44.0
wrapt==1.16.0
zarr==2.18.3
